requirements.txt
python-dotenv==1.1.1
google-auth==2.40.3
google-cloud-bigquery==3.36.0
google-cloud-discoveryengine==0.13.11
google-cloud-storage==3.3.0
pydantic==2.11.7
python-dotenv==1.1.1
pytest==8.4.1
invoke==2.2.0
